prometheus, Add prometheus-related instances
Deploy CNAO's prometheus-related instances on the cluster
(serviceMonitor, PrometheusRules, rbacs, etc..), to serve components
and collect metrics according to the PrometheusRule instance.
Will only deploy if monitoring namespace is created, which is
configurable as CNAO container variable

Note: These monitoring instances are treated as a CNAO
component, and as such get the operator reconcile & labeling services.

Signed-off-by: Ram Lavi <ralavi@redhat.com>