deployment tests: Increase timeout waiting for Progressing state
We now deploy cert-manager and ipam controller,
the ipam controller uses a secret created by cert-manager,
it takes time until cert-manager is deployed and the secret is created,
in this transient time, the state might be Failed but eventually
turn to Progressing and finally Available.
Increate timeout because of this.

Signed-off-by: Or Shoval <oshoval@redhat.com>